Add add separate parameters module
